+++ src/decorator/AsyncMethodDecorator.ts
@@ -1,13 +1,13 @@
 export type VoidFunc = (() => Promise<void>) | (() => void);
 export type ErrorArgVoidFunc = ((error: any) => Promise<void>) | (() => void);
 /**
- * A base of Decorator for using async/await
+ * A template of Method Decorator for using async/await
  * @param beforeFunc no args void func
  * @param afterFunc no args void func
  * @param catchFunc error arg void func
  * @param finalFunc no args void func
  */
-export function AsyncDecorator(
+export function AsyncMethodDecorator(
   beforeFunc?: VoidFunc,
   afterFunc?: VoidFunc,
   catchFunc?: ErrorArgVoidFunc,
